pyproject.toml
[tool.poetry]
name = "orchestration-play"
version = "0.1.0"
description = "Orchestration flows with Prefect"
authors = ["Khalid MAJDOUB <khalid.majdoub@sidetrade.com>"]
packages = [
    { include = "ccfarm", from = "src" },
    { include = "common", from = "src" },
    { include = "draft", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
requests = ">=2.31.0"
anthropic = ">=0.3.11"
python-dotenv = ">=1.0.0"
tenacity = "^9.0.0"
trafilatura = "^2.0.0"
bs4 = "^0.0.2"
pymongo = "^4.11.1"
autogen = "<0.7.0"
tabulate = "^0.9.0"
soundfile = "^0.13.1"
pydub = "^0.25.1"
boto3 = "^1.37.9"
prefect = "3.1.3"
prefect-aws = "^0.5.5"
elevenlabs = "^1.54.0"
moviepy = "^2.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.10.0"
types-requests = "^2.32.0.20241016"
types-python-dateutil = "^2.9.0.20241206"
types-pyyaml = "^6.0.12.20241230"
types-beautifulsoup4 = "^4.12.0.20250204"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python = ">=3.10,<3.14"
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
disallow_untyped_defs = false
follow_imports = "silent"
strict_optional = false
disable_error_code = ["attr-defined"]

[tool.pyright]
reportOptionalMemberAccess = false
reportOptionalCall = false
reportOptionalIterable = false
reportOptionalContextManager = false
reportOptionalOperand = false